Feature: CMake Presets (#39)
* update gitignore
* Add cmake presets
* Properly set openmp compile and link options
* Add cmake presets to CI
* Add cmake presets to tests
* Create .dSYM directories for debug info on macOS when doing debug builds
* Bump rapids-cmake to 23.10
* Set Legion as system in its configure stage
* Mark Python headers as internal for cython builds
* Centralize setting legate_core_ROOT if not needed
* Always download the correct rapids-cmake version
* Fix warnings as a result of enabling -Werror

+++ src/core/runtime/detail/partition_manager.cc
@@ -141,7 +141,7 @@ Shape PartitionManager::compute_launch_shape(const mapping::Machine& machine,
     size_t factor_prod = 1;
     for (auto factor : get_factors(machine)) {
       // Avoid exceeding the maximum number of pieces
-      if (factor * factor_prod > max_pieces) break;
+      if (factor * factor_prod > static_cast<std::size_t>(max_pieces)) break;
 
       factor_prod *= factor;
